Use new caching base for CachedPoints/Content
Prepares for further modifications.
