+++ exercise/api/csv/views.py
@@ -105,7 +105,7 @@ class CourseSubmissionDataViewSet(NestedViewSetMixin,
         # is visible to the request user, and not the points for that user.
         # Therefore it is okay to use CachedPoints, even though this view
         # includes many users and CachedPoints includes only one.
-        points = CachedPoints(self.instance, request.user, self.content, self.is_course_staff)
+        points = CachedPoints(self.instance, request.user, self.is_course_staff)
         ids = [e.id for e in self.content.search_exercises(**search_args)]
         revealed_ids = get_revealed_exercise_ids(search_args, points)
         queryset = Submission.objects.filter(
@@ -123,7 +123,7 @@ class CourseSubmissionDataViewSet(NestedViewSetMixin,
             ) -> Response:
         profile = self.get_object()
         search_args = self.get_search_args(request)
-        points = CachedPoints(self.instance, profile.user, self.content, self.is_course_staff)
+        points = CachedPoints(self.instance, profile.user, self.is_course_staff)
         ids = points.submission_ids(**search_args)
         revealed_ids = get_revealed_exercise_ids(search_args, points)
         queryset = Submission.objects.filter(
@@ -234,7 +234,7 @@ class CourseAggregateDataViewSet(NestedViewSetMixin,
         search_args = self.get_search_args(request)
         entry, exercises = self.content.search_entries(**search_args)
         ids = [e.id for e in exercises if e.type == 'exercise']
-        points = CachedPoints(self.instance, request.user, self.content, self.is_course_staff)
+        points = CachedPoints(self.instance, request.user, self.is_course_staff)
         revealed_ids = get_revealed_exercise_ids(search_args, points)
         aggr = (
             Submission.objects
@@ -411,7 +411,7 @@ class CourseResultsDataViewSet(NestedViewSetMixin,
         search_args = self.get_search_args(request)
         exercises = self.content.search_exercises(**search_args)
         ids = [e.id for e in exercises]
-        points = CachedPoints(self.instance, request.user, self.content, self.is_course_staff)
+        points = CachedPoints(self.instance, request.user, self.is_course_staff)
         revealed_ids = get_revealed_exercise_ids(search_args, points)
         exclude_list = [Submission.STATUS.ERROR, Submission.STATUS.REJECTED]
         show_unofficial = request.GET.get('show_unofficial') == 'true'
