+++ exercise/cache/basetypes.py
@@ -1,7 +1,20 @@
 from __future__ import annotations
 from dataclasses import dataclass, field
 from datetime import datetime
-from typing import Any, ClassVar, Dict, Generic, List, Literal, Optional, TypeVar
+from typing import Any, ClassVar, Dict, Generic, List, Literal, Optional, TypeVar, Union
+
+from django.utils import timezone
+
+from course.models import CourseInstance
+from lib.cache.cached import CacheBase
+from ..models import BaseExercise, LearningObject
+
+
+def add_by_difficulty(to: Dict[str, int], difficulty: str, points: int):
+    if difficulty in to:
+        to[difficulty] += points
+    else:
+        to[difficulty] = points
 
 
 class EqById:
@@ -105,7 +118,8 @@ class TotalsBase:
 
 
 @dataclass
-class CachedDataBase(Generic[ModuleEntry, ExerciseEntry, CategoryEntry, Totals]):
+class CachedDataBase(CacheBase, Generic[ModuleEntry, ExerciseEntry, CategoryEntry, Totals]):
+    KEY_PREFIX: ClassVar[str] = 'instance'
     created: datetime
     module_index: Dict[int, ModuleEntry]
     exercise_index: Dict[int, ExerciseEntry]
@@ -113,3 +127,155 @@ class CachedDataBase(Generic[ModuleEntry, ExerciseEntry, CategoryEntry, Totals])
     modules: List[ModuleEntry]
     categories: Dict[int, CategoryEntry]
     total: Totals
+
+    # pylint: disable-next=arguments-differ too-many-locals
+    @classmethod
+    def _generate_data(
+            cls,
+            instance_id: int,
+            ) -> CachedDataBase[ModuleEntryBase, ExerciseEntryBase, CategoryEntryBase, TotalsBase]:
+        """ Returns object that is cached into self.data """
+        instance = CourseInstance.objects.get(id=instance_id)
+
+        exercise_index: Dict[int, ExerciseEntryBase] = {}
+        module_index: Dict[int, ModuleEntryBase] = {}
+        paths: Dict[int, Dict[str, int]] = {}
+        modules: List[ModuleEntryBase] = []
+        categories: Dict[int, CategoryEntryBase] = {}
+        total = TotalsBase()
+
+        def recursion(
+                module: ModuleEntryBase,
+                objects: List[LearningObject],
+                parents: List[LearningObject],
+                container: List[ExerciseEntryBase],
+                ) -> None:
+            """ Recursively travels exercises hierarchy """
+            parent_id = parents[-1].id if parents else None
+            children = [o for o in objects if o.parent_id == parent_id]
+            j = 0
+            for o in children:
+                o._parents = parents + [o]
+                category = o.category
+                entry = ExerciseEntryBase(
+                    module = module,
+                    parent = exercise_index[parent_id] if parent_id is not None else None,
+                    category = str(category),
+                    category_id = category.id,
+                    category_status = category.status,
+                    confirm_the_level = category.confirm_the_level,
+                    module_id = module.id,
+                    module_status = module.status,
+                    id = o.id,
+                    order = o.order,
+                    status = o.status,
+                    name = str(o),
+                    hierarchical_name = o.hierarchical_name(),
+                    number = module.number + '.' + o.number(),
+                    link = o.get_display_url(),
+                    submittable = False,
+                    submissions_link = o.get_submission_list_url(),
+                    requirements = module.requirements,
+                    opening_time = module.opening_time,
+                    reading_opening_time = module.reading_opening_time,
+                    closing_time = module.closing_time,
+                    late_allowed = module.late_allowed,
+                    late_time = module.late_time,
+                    late_percent = module.late_percent,
+                    is_empty = o.is_empty(),
+                    get_path = o.get_path(),
+                )
+                container.append(entry)
+                exercise_index[o.id] = entry
+                paths[module.id][o.get_path()] = o.id
+                if category.id not in categories:
+                    categories[category.id] = CategoryEntryBase(
+                        id = category.id,
+                        status = category.status,
+                        name = str(category),
+                        points_to_pass = category.points_to_pass,
+                    )
+                recursion(module, objects, o._parents, entry.children)
+                j += 1
+
+        # Collect each module.
+        i = 0
+        for module in instance.course_modules.prefetch_related(
+            'requirements',
+            'requirements__threshold__passed_modules',
+            'requirements__threshold__passed_categories',
+            'requirements__threshold__passed_exercises',
+            'requirements__threshold__passed_exercises__parent',
+            'requirements__threshold__points',
+            'learning_objects',
+        ):
+            entry = ModuleEntryBase(
+                id = module.id,
+                order = module.order,
+                status = module.status,
+                url = module.url,
+                name = str(module),
+                number = str(module.order),
+                introduction = module.introduction,
+                link = module.get_absolute_url(),
+                requirements = [str(r) for r in module.requirements.all()],
+                opening_time = module.opening_time,
+                reading_opening_time = module.reading_opening_time,
+                closing_time = module.closing_time,
+                late_allowed = module.late_submissions_allowed,
+                late_time = module.late_submission_deadline,
+                late_percent = module.get_late_submission_point_worth(),
+                points_to_pass = module.points_to_pass,
+            )
+            modules.append(entry)
+            module_index[module.id] = entry
+            paths[module.id] = {}
+            all_children = list(module.learning_objects.all())
+            recursion(entry, all_children, [], entry.children)
+            i += 1
+
+        # Augment submittable exercise parameters.
+        def add_to(target: Union[ModuleEntryBase, CategoryEntryBase, TotalsBase], exercise: BaseExercise) -> None:
+            target.exercise_count += 1
+            target.max_points += exercise.max_points
+            add_by_difficulty(
+                target.max_points_by_difficulty,
+                exercise.difficulty,
+                exercise.max_points
+            )
+        for exercise in BaseExercise.objects\
+              .filter(course_module__course_instance=instance):
+            try:
+                entry = exercise_index[exercise.id]
+            except KeyError:
+                continue
+
+            entry.submittable = True
+            entry.points_to_pass = exercise.points_to_pass
+            entry.difficulty = exercise.difficulty
+            entry.max_submissions = exercise.max_submissions
+            entry.max_points = exercise.max_points
+            entry.allow_assistant_viewing = exercise.allow_assistant_viewing
+
+            if not entry.confirm_the_level:
+                add_to(entry.module, exercise)
+                add_to(categories[exercise.category.id], exercise)
+                add_to(total, exercise)
+
+                if exercise.max_group_size > total.max_group_size:
+                    total.max_group_size = exercise.max_group_size
+                if exercise.max_group_size > 1 and exercise.min_group_size < total.min_group_size:
+                    total.min_group_size = exercise.min_group_size
+
+        if total.min_group_size > total.max_group_size:
+            total.min_group_size = 1
+
+        return CachedDataBase(
+            created = timezone.now(),
+            module_index = module_index,
+            exercise_index = exercise_index,
+            paths = paths,
+            modules = modules,
+            categories = categories,
+            total = total,
+        )
