+++ exercise/staff_views.py
@@ -139,7 +139,7 @@ class InspectSubmitterView(ExerciseBaseView, BaseRedirectView):
         )
 
         # Find the submitter's best submission using the cache.
-        cache = CachedPoints(self.instance, user, self.content, True)
+        cache = CachedPoints(self.instance, user, True)
         ids = cache.submission_ids(exercise_id=self.exercise.id, best=True, fallback_to_last=True)
         if not ids:
             raise Http404()
@@ -352,7 +352,7 @@ class NextUnassessedSubmitterView(ExerciseBaseView, BaseRedirectView):
             return self.redirect(self.exercise.get_submission_list_url())
 
         # Find the submitter's best submission using the cache.
-        cache = CachedPoints(self.instance, submitter.user, self.content, True)
+        cache = CachedPoints(self.instance, submitter.user, True)
         ids = cache.submission_ids(exercise_id=self.exercise.id, best=True, fallback_to_last=True)
         if not ids:
             raise Http404()
