+++ lib/cache/__init__.py
@@ -1,2 +1,2 @@
 from .backends import _LocMemCache
-from .cached import CachedAbstract
+from .cached_old import CachedAbstract
